+++ webtack/jsinterop-generator/src/test/fixtures/input_catalog/output/gwt/main/java/com/example/IndexedKeyframeOffsetType.java
@@ -29,7 +29,7 @@ public interface IndexedKeyframeOffsetType {
 
   @JsOverlay
   @Nonnull
-  static IndexedKeyframeOffsetType of(@Nonnull final double[] value) {
+  static IndexedKeyframeOffsetType of(@Nonnull final double... value) {
     return Js.cast( value );
   }