+++ webtack/jsinterop-generator/src/test/fixtures/js_predefined_symbols/output/gwt/main/java/com/example/Int32List2.java
@@ -28,7 +28,7 @@ public interface Int32List2 {
 
   @JsOverlay
   @Nonnull
-  static Int32List2 of(@Nonnull final double[] value) {
+  static Int32List2 of(@Nonnull final double... value) {
     return Js.cast( value );
   }