+++ webtack/jsinterop-generator/src/test/fixtures/js_predefined_symbols/output/j2cl/main/java/com/example/Int32List.java
@@ -28,7 +28,7 @@ public interface Int32List {
 
   @JsOverlay
   @Nonnull
-  static Int32List of(@Nonnull final double[] value) {
+  static Int32List of(@Nonnull final double... value) {
     return Js.cast( value );
   }