+++ webtack/jsinterop-generator/src/test/fixtures/optional_guards/output/j2cl/main/java/com/example/VibratePattern.java
@@ -28,7 +28,7 @@ public interface VibratePattern {
 
   @JsOverlay
   @Nonnull
-  static VibratePattern of(@Nonnull final double[] value) {
+  static VibratePattern of(@Nonnull final double... value) {
     return Js.cast( value );
   }