+++ webtack/jsinterop-generator/src/test/fixtures/unions/output/j2cl/main/java/com/example/StringOrStringArrayUnion.java
@@ -28,7 +28,7 @@ public interface StringOrStringArrayUnion {
 
   @JsOverlay
   @Nonnull
-  static StringOrStringArrayUnion of(@Nonnull final String[] value) {
+  static StringOrStringArrayUnion of(@Nonnull final String... value) {
     return Js.cast( value );
   }