+++ webtack/jsinterop-generator/src/test/fixtures/unions/output/j2cl/main/java/com/example/UnionBooleanSequenceOrString.java
@@ -22,7 +22,7 @@ public interface UnionBooleanSequenceOrString {
 
   @JsOverlay
   @Nonnull
-  static UnionBooleanSequenceOrString of(@Nonnull final Boolean[] value) {
+  static UnionBooleanSequenceOrString of(@Nonnull final Boolean... value) {
     return Js.cast( value );
   }