+++ webtack/jsinterop-generator/src/test/fixtures/unions/output/j2cl/main/java/com/example/UnionDoubleSequenceOrString.java
@@ -22,7 +22,7 @@ public interface UnionDoubleSequenceOrString {
 
   @JsOverlay
   @Nonnull
-  static UnionDoubleSequenceOrString of(@Nonnull final double[] value) {
+  static UnionDoubleSequenceOrString of(@Nonnull final double... value) {
     return Js.cast( value );
   }